fix: Fix service shutdown (#5)
Sometimes, Gource does not get killed, causing the service shutdown to
hang for a few minutes. Setting KillMode=mixed in the service definition
tells systemd to send SIGTERM to the main process, then SIGKILL (which
cannot be ignored) to all other processes in the group. This cleans up
Gource right away.